pyproject.toml
[project]
name = "deepflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click==8.1.8",
    "graphviz>=0.21",
    "joblib>=1.5.2",
    "numpy==2.3.3",
    "pandas==2.3.1",
    "protobuf==5.28.2",
    "pydot>=3.0.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "ruamel-yaml==0.17.40",
    "ruamel-yaml-clib==0.2.8",
    "seaborn>=0.13.2",
    "six==1.17.0",
    "tzdata==2025.2",
]
